added deduct credit at start of startProject event and restored in catch block of processConversation
