+++ server/controllers/onboardingToolController.js
@@ -4,13 +4,13 @@ const searchService = require("../services/searchService");
 const { toKebabCase } = require("../utils/file");
 const {
   insertShip,
-  getUserProfile,
-  updateUserProfile,
+  // getUserProfile,
+  // updateUserProfile,
   updateConversation,
 } = require("../services/dbService");
 const { TOOLS } = require("../config/tools");
 
-const { nanoid } = require('nanoid');
+const { nanoid } = require("nanoid");
 
 const generateProjectFolderName = (projectName) => {
   return toKebabCase(projectName) + "-" + nanoid(8);
@@ -145,12 +145,14 @@ async function handleOnboardingToolUse({
     };
     const { id } = await insertShip(ship);
     console.log("Inserted ship", id);
-    if (mode === 'paid') {
-      const profile = await getUserProfile(userId);
-      const { available_ships } = profile; // current
-      const profilePayload = { available_ships: available_ships - 1 }; // updated
-      await updateUserProfile(userId, profilePayload);
-    }
+
+    // removing deduct available ships from here.
+    // if (mode === 'paid') {
+    //   const profile = await getUserProfile(userId);
+    //   const { available_ships } = profile; // current
+    //   const profilePayload = { available_ships: available_ships - 1 }; // updated
+    //   await updateUserProfile(userId, profilePayload);
+    // }
     const convPayload = {
       ship_id: id,
       tokens_used: client.tokensUsed,
