+++ server/services/onboadingService.js
@@ -13,7 +13,11 @@ const {
   handleOnboardingToolUse,
 } = require("../controllers/onboardingToolController");
 const { AnthropicService } = require("../services/anthropicService");
-const { getUserProfile, insertMessage } = require("../services/dbService");
+const {
+  getUserProfile,
+  insertMessage,
+  updateUserProfile,
+} = require("../services/dbService");
 const { SHIP_TYPES, DEFAULT_MESSAGES } = require("./constants");
 
 async function processConversation({
@@ -52,7 +56,6 @@ async function processConversation({
         tools.push(searchTool);
         messages = [{ role: "user", content: message }];
       }
-
     }
 
     try {
@@ -188,7 +191,14 @@ function handleOnboardingSocketEvents(io) {
       };
 
       abortController = new AbortController();
+      const mode = client.isCustomKey ? "self-key" : "paid";
       try {
+        if (mode === "paid") {
+          const profile = await getUserProfile(userId);
+          const { available_ships } = profile; // current
+          const profilePayload = { available_ships: available_ships - 1 }; // updated
+          await updateUserProfile(userId, profilePayload);
+        }
         await processConversation({
           client,
           sendEvent,
@@ -206,6 +216,14 @@ function handleOnboardingSocketEvents(io) {
           sendEvent("creationAborted", {
             message: "Website creation was aborted",
           });
+
+          // revert available ships in case of paid mode
+          if (mode === "paid") {
+            const profile = await getUserProfile(userId);
+            const { available_ships } = profile;
+            const profilePayload = { available_ships: available_ships + 1 }; // updated
+            await updateUserProfile(userId, profilePayload);
+          }
         } else {
           console.error("Error in processConversation:", error);
         }
